+++ src/session.h
@@ -15,6 +15,8 @@
 #ifndef SESSION_H
 #define SESSION_H
 
-int session_new_connection(const char *method, const char *url);
+#include <stdint.h>
+
+uint64_t session_new_connection(const char *method, const char *url);
 
 #endif
